refactor: improve DeviceInfoProvider implementation and location
