+++ app/src/main/java/com/hilingual/di/ProviderModule.kt
@@ -1,7 +1,7 @@
 package com.hilingual.di
 
 import com.hilingual.core.common.provider.DeviceInfoProvider
-import com.hilingual.provider.AndroidDeviceInfoProvider
+import com.hilingual.app.DeviceInfoProviderImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -15,6 +15,6 @@ internal abstract class ProviderModule {
     @Binds
     @Singleton
     abstract fun bindDeviceInfoProvider(
-        impl: AndroidDeviceInfoProvider
+        impl: DeviceInfoProviderImpl
     ): DeviceInfoProvider
 }
